move non user facing IGraphics structs to IGraphicsPrivate.h
